Fix Output Value Class of AverageComputationAlgorithm

+++ PROJECT PARTS 1 - 5/AverageComputationAlgorithm/src/net/endoedgar/AverageComputationAlgorithm.java
@@ -17,7 +17,7 @@ public class AverageComputationAlgorithm {
 		job.setJarByClass(AverageComputationAlgorithm.class);
 		
 		job.setOutputKeyClass(Text.class);
-		job.setOutputValueClass(IntWritable.class);
+		job.setOutputValueClass(DoubleWritable.class);
 		
 		job.setMapperClass(MyMapper.class);
 		job.setReducerClass(MyReducer.class);
